@@ -107,8 +107,8 @@
 					<groupId>org.apache.maven.plugins</groupId>
 					<artifactId>maven-compiler-plugin</artifactId>
 					<configuration>
-						<target>8</target>
-						<source>8</source>
+						<target>11</target>
+						<source>11</source>
 					</configuration>
 				</plugin>
 				<plugin>
